Add obs.eval, fix bug, add unit test, apply comments
